Reduce items in delegates
We have a lot of delegates, we need them as light as possible.

Currently a simple cell delegate contains:
 - A mouseArea for tooltips
 - A QQuickToolTipAttached (albeit, this is just a light object)
 - Another mouseArea in the background handling clicks

We can move this all up to be one MouseArea covering the table. It
requires a tiny bit more manual mapping, and the view making assumptions
about the delegates, but it's a big reduction in objects.